Rename ExplanationTop into ElExplanationHeader
